more refactors out
