+++ src/aioway/hop/__init__.py
@@ -1,7 +1,6 @@
 # Copyright (c) AIoWay Authors - All Rights Reserved
 
 from .funcs import *
-from .hop import *
 from .iters import *
 from .joins import *
 from .loaders import *
