--- src/aioway/hop/hop.py
@@ -1,313 +1,0 @@
-# Copyright (c) AIoWay Authors - All Rights Reserved
-
-"The [h]igh level [o]peration [p]review class."
-
-import abc
-import copy
-import dataclasses as dcls
-import textwrap
-import typing
-from collections import abc as cabc
-
-import tensordict as td
-import torch
-from rich import table
-
-from aioway._utils import (
-    AnyDict,
-    AnySet,
-    dag_node_key,
-    dcls_asdict,
-    decomp_dcls_members,
-    decomp_replace,
-    render_fcall_str,
-    topo_sort,
-    torch_fake_mode,
-)
-from aioway.spaces import Attr, Shape
-
-if typing.TYPE_CHECKING:
-    from aioway.hop import IterProc
-
-__all__ = ["iter_dcls", "Iter", "TensorIter", "TdictIter", "ListIter", "BoundedIter"]
-
-
-@typing.dataclass_transform()
-def iter_dcls(cls: type):
-    return dcls.dataclass(match_args=False)(cls)
-
-
-@iter_dcls
-class Iter[T](cabc.Iterable[T], abc.ABC):
-    """
-    The class that defines [h]igh level [op]erations.
-    It produces iterators that computes the desired batch, represented by the node.
-
-    `Iter` is the node that would be evaluated during run time.
-    It will output `torch.Tensor` a container that makes up of them.
-    """
-
-    def __hash__(self) -> int:
-        return id(self)
-
-    @typing.final
-    def __iter__(self) -> IterProc[T]:
-        # Every iteration should yield a new `Iterator`.
-        from .iters import IterProc
-
-        return IterProc(self)
-
-    @abc.abstractmethod
-    def iterate(self) -> cabc.Iterator[T]:
-        """
-        The iteration logic.
-        Should invoke dependencies' via `__iter__` methods (just use `for` loops).
-        """
-
-        raise NotImplementedError
-
-    def sample(self) -> T:
-        """
-        Get a sample from the input. This method should be equivalent to `next(iter(self))`,
-        but in fake mode (no data is actually retrieved).
-        """
-
-        with torch_fake_mode():
-            return next(iter(self))
-
-    @property
-    def requires_grad(self) -> bool:
-        """
-        Whether or not the `torch.Tensor`s generated by this `Iter` `requires_grad`.
-        If `False`, much like `torch`'s `requires_grad`,
-        gradient does not need to propagate to this node, and we can do a graph break.
-
-        Defaults to checking if any of the `self.deps()` have any `requires_grad`.
-        """
-
-        return any(hop.requires_grad for hop in self.deps())
-
-    def replace(
-        self,
-        function: cabc.Callable[[Iter], Iter],
-        memo: AnyDict[typing.Any, typing.Any] | None = None,
-    ) -> typing.Self:
-        """
-        Replace the current `node` with `function(node)` if it's not `None`,
-        or else recursively invoke `.replace(function)` on `Iter` in `.inputs()`.
-
-        Args:
-            function: The function to apply to the graph.
-            memo: The memo to use if given.
-        """
-
-        def maybe_replace(hop):
-            if not isinstance(hop, Iter):
-                return NotImplemented
-
-            return function(hop)
-
-        result: typing.Any = decomp_replace(self, maybe_replace, memo=memo)
-        assert type(result) == type(self)
-        return result
-
-    def rebuild(self):
-        """
-        Rebuild the current `Iter`. This is useful when you are switching contexts,
-        e.g. switching on real mode after configuring the `HopDag` in fake mode.
-
-        If `self._rebuild()` is not overwritten, defaults to shallow copying `self`.
-        """
-
-        copied = self._rebuild()
-        assert copied is not self
-        return copied
-
-    def _rebuild(self) -> typing.Self:
-        return copy.copy(self)
-
-    def deps(self) -> cabc.Iterator[Iter]:
-        "Decompose `self`, get the immediate dependencies."
-
-        for hop in decomp_dcls_members(self, Iter):
-            yield hop
-
-    def nodes(self) -> AnySet[Iter]:
-        "Get all the nodes recursively, including `self`."
-
-        nodes = AnySet[Iter](Iter)
-
-        # Add nodes recursively.
-        def _visit_nodes(hop: Iter):
-            if hop in nodes:
-                return
-
-            nodes.add(hop)
-
-            for dep in hop.deps():
-                _visit_nodes(dep)
-
-        _visit_nodes(self)
-        return nodes
-
-    @property
-    def is_source(self) -> bool:
-        "If the `Iter` is not dependent on other `Iter`s, it's considered a source."
-
-        for _ in decomp_dcls_members(self, Iter):
-            return True
-        else:
-            return False
-
-    @property
-    def size(self) -> int:
-        """
-        The length of the current stream `TdictIter`.
-
-        This should be defined for relational algebra purposes.
-        """
-
-        return NotImplemented
-
-    def dag(self) -> list[Iter[T]]:
-        "Order the `Iter`s in their topological order."
-
-        dag_nodes = [dag_node_key(hop) for hop in self.nodes()]
-        return topo_sort(dag_nodes)
-
-    def _repr_abbrev(self) -> str:
-        """
-        Perform `repr`, but with `Iter` members abbreviated.
-        """
-
-        members: dict[str, str] = {}
-
-        for key, val in dcls_asdict(self).items():
-
-            # Only perform abbreviation on `Iter`.
-            if isinstance(val, Iter):
-                string = f"{type(val).__name__}(...)"
-            else:
-                string = textwrap.shorten(repr(val), 20)
-
-            members[key] = string
-
-        return render_fcall_str(type(self).__name__, **members)
-
-    def explain(self) -> table.Table:
-        tab = table.Table(title="Logical plan")
-        tab.add_column("Id")
-        tab.add_column("Parents")
-        tab.add_column("Self")
-
-        dag = self.dag()
-        hop_idx = AnyDict[Iter, int]()
-        for idx, hop in enumerate(dag):
-            hop_idx[hop] = idx
-
-        for idx, hop in enumerate(dag):
-            parents = (f"#{hop_idx[dep]}" for dep in hop.deps())
-            tab.add_row(f"#{idx}", ",".join(parents), hop._repr_abbrev())
-
-        return tab
-
-
-@iter_dcls
-class TensorIter(Iter[torch.Tensor], abc.ABC):
-    """
-    An iterator of batches of `torch.Tensor`.
-    """
-
-    @property
-    def attr(self) -> Attr:
-        return Attr.parse(self.sample())
-
-    @property
-    def shape(self) -> Shape:
-        return self.attr.shape
-
-    @property
-    def ndim(self) -> int:
-        return self.attr.ndim
-
-
-@iter_dcls
-class TdictIter(Iter[td.TensorDict], abc.ABC):
-    """
-    An iterator of batches of `td.TensorDict`.
-
-    This is the core abstraction used by the relational algebra operators.
-    """
-
-    def column(self, col: str) -> TensorIter:
-        from .views import ColumnViewIter
-
-        return ColumnViewIter(self, col)
-
-    def select(self, *cols: str) -> TdictIter:
-        from .views import ProjectIter
-
-        return ProjectIter(self, subset=list(cols))
-
-
-@iter_dcls
-class ListIter[T = typing.Any](Iter[cabc.Sequence[T]]):
-    "A convenient list of `Iter`s, using a pull strategy to pull in the data when called."
-
-    hops: cabc.Sequence[Iter[T]]
-    """
-    The hop list that this `HopList` represents.
-    """
-
-    def __repr__(self) -> str:
-        return repr(self.hops)
-
-    def __len__(self) -> int:
-        return len(self.hops)
-
-    @typing.overload
-    def __getitem__(self, key: int) -> Iter[T]: ...
-
-    @typing.overload
-    def __getitem__(self, key: slice) -> typing.Self: ...
-
-    def __getitem__(self, key):
-        if isinstance(key, int):
-            return self.hops[key]
-
-        if isinstance(key, slice):
-            return type(self)(self.hops[key])
-
-        raise TypeError(f"Don't know how to handle {type(key)=}.")
-
-    @typing.override
-    def iterate(self):
-        for hops in zip(*self.hops):
-            yield hops
-
-
-@iter_dcls
-class BoundedIter(TdictIter, abc.ABC):
-    """
-    A stream with `__len__` and `__getitem__`.
-    """
-
-    @abc.abstractmethod
-    def __len__(self) -> int:
-        "The number of batches saved in the current `Stream`."
-
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def __getitem__(self, key: int, /) -> td.TensorDict:
-        """
-        Get individual items. Does not support slice input.
-
-        Args:
-            idx: An integer. Must be in the range `[-len(self), len(self))`.
-
-        Returns:
-            The `td.TensorDict` batch.
-        """
-
-        raise NotImplementedError
