+++ src/aioway/hop/maps.py
@@ -10,9 +10,8 @@ from collections import abc as cabc
 import tensordict as td
 import torch
 
+from aioway._comps import TdictIter, iter_dcls
 from aioway._utils import tdict_rename
-
-from .hop import TdictIter, iter_dcls
 
 __all__ = ["MapIter", "ApplyIter", "FuncFilterIter", "RenameIter"]
