+++ src/aioway/indices.py
@@ -9,8 +9,8 @@ from collections import abc as cabc
 import numpy as np
 import torch
 
+from aioway._comps import Iter, TensorIter, iter_dcls
 from aioway._utils import FloatArray, IntArray
-from aioway.hop import Iter, TensorIter, iter_dcls
 
 __all__ = ["AnnIndex", "AnnResult", "AnnIndexIter", "AnnIndexTrainerIter", "FaissIndex"]
