+++ tests/hop/test_binary.py
@@ -7,9 +7,10 @@ import pytest
 import tensordict as td
 import torch
 
+from aioway._comps import TdictIter
 from aioway._utils import tdict_all_equal
 from aioway.dsets import TdictListIter
-from aioway.hop import NestedLoopJoinIter, TdictIter, ZipIter
+from aioway.hop import NestedLoopJoinIter, ZipIter
 
 
 @pytest.fixture
