+++ tests/torch/nn/test_hop.py
@@ -6,9 +6,10 @@ import pytest
 import torch
 from torch import nn
 
+from aioway._comps import Iter, TensorIter
 from aioway._utils import AnyDict, is_fake_tensor, torch_fake_mode
 from aioway.dsets import TensorListIter
-from aioway.hop import Iter, StackIter, TensorIter, ufunc_cache_on
+from aioway.hop import StackIter, ufunc_cache_on
 from aioway.modes import NnInitThunk
 from aioway.torch.nn import NnLayerIter, NnLossIter, build_nn_hop
